requirements.txt
fastapi==0.115.13
langchain==0.3.25
langchain_community==0.3.25
langchain_core==0.3.65
langchain_google_genai==2.1.5
langchain_openai==0.3.24
langchain_text_splitters==0.3.8
logfire==3.21.1
mpxpy==0.0.15
Pillow==11.2.1
pydantic==2.11.7
python-dotenv==1.1.0
PyYAML==6.0.2
Requests==2.32.4
streamlit==1.46.0
tqdm==4.67.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.34.3
